feature(Cart): componentizacioo de CartItemList y PurchaseDetail en un componente padre Cart
